criado component para armazenagem de pedidos
